+++ backend/internal/services/calcutta_service.go
@@ -83,7 +83,7 @@ func (s *CalcuttaService) CalculateOwnershipPercentage(team *models.CalcuttaEntr
 }
 
 // CalculatePoints calculates the points earned by a team based on its performance
-func (s *CalcuttaService) CalculatePoints(team *models.TournamentTeam, rounds []*models.CalcuttaRound) int {
+func (s *CalcuttaService) CalculatePoints(team *models.TournamentTeam, rounds []*models.CalcuttaRound) float64 {
 	// Calculate total progress (wins + byes)
 	totalProgress := team.Wins + team.Byes
 
@@ -120,8 +120,8 @@ func (s *CalcuttaService) CalculatePoints(team *models.TournamentTeam, rounds []
 }
 
 // CalculatePlayerPoints calculates the points earned by a player based on their team ownerships
-func (s *CalcuttaService) CalculatePlayerPoints(portfolio *models.CalcuttaPortfolio, portfolioTeams []*models.CalcuttaPortfolioTeam) int {
-	totalPoints := 0
+func (s *CalcuttaService) CalculatePlayerPoints(portfolio *models.CalcuttaPortfolio, portfolioTeams []*models.CalcuttaPortfolioTeam) float64 {
+	totalPoints := 0.0
 	for _, portfolioTeam := range portfolioTeams {
 		if portfolioTeam.PortfolioID == portfolio.ID {
 			totalPoints += portfolioTeam.PointsEarned
